refactor clase test para extraer metodos menu
